pyproject.toml
[project]
name = "GrantBox"
version = "0.1.0"
description = "Agent execution sandbox with LangChain/LangGraph, MCP tools, and defense hooks"
authors = [{ name = "", email = "" }]
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "langchain>=0.3.0",
  "langgraph>=0.2.21",
  "langchain-openai>=0.2.0",
  "langchain-google-genai>=2.0.0",
  "langchain-anthropic>=0.2.0",
  "langchain-community>=0.3.0",
  "httpx[socks]>=0.27",
  "tenacity>=8.2",
  "dashscope>=1.24.9",
  "python-dotenv>=1.2.1",
  "matplotlib>=3.10.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "ruff>=0.6.9",
  "mypy>=1.10",
]

[tool.uv]
index-url = "https://pypi.org/simple"

 
